luz-py/requirements.txt
numpy==1.24.3
torch==2.2.2
pillow==11.1.0
transformers==4.48.2
huggingface-hub==0.28.1
requests==2.32.3
urllib3==2.3.0
idna==3.10
certifi==2025.1.31
tqdm==4.67.1
PyYAML==6.0.2
filelock==3.17.0
packaging==24.2
typing-extensions==4.12.2
regex==2024.11.6
safetensors==0.5.2
tokenizers==0.21.0
